fix code and tests
